backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
pydantic==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sse-starlette==2.1.2
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
openpyxl==3.1.5
python-multipart==0.0.6
